Complete restructuring of API calls
