+++ IMDB_API_Accessor/API/IMDBAPI.cs
@@ -8,7 +8,7 @@ namespace APIAccessor.API
     public class IMDBAPI : APIWrapper<IMDBMetaData>
     {
         public IMDBAPI(String authKey) : 
-            base(authKey, "https://movie-database-imdb-alternative.p.rapidapi.com/?i={id}&r=json")
+            base(authKey, "https://imdb8.p.rapidapi.com/title/get-meta-data?region=US&ids={0}")
         {
         }
